break up browser base into filter mixins
